Align numeric types with spec

+++ src/main/java/com/amannmalik/mcp/transport/StreamableHttpTransport.java
@@ -159,7 +159,7 @@ public final class StreamableHttpTransport implements Transport {
                         reqId = new RequestId.StringId(id.substring(1, id.length() - 1));
                     } else {
                         try {
-                            reqId = new RequestId.NumericId(Double.parseDouble(id));
+                            reqId = new RequestId.NumericId(Long.parseLong(id));
                         } catch (NumberFormatException e) {
                             reqId = new RequestId.StringId(id);
                         }
@@ -183,7 +183,7 @@ public final class StreamableHttpTransport implements Transport {
                     reqId = new RequestId.StringId(id.substring(1, id.length() - 1));
                 } else {
                     try {
-                        reqId = new RequestId.NumericId(Double.parseDouble(id));
+                        reqId = new RequestId.NumericId(Long.parseLong(id));
                     } catch (NumberFormatException e) {
                         reqId = new RequestId.StringId(id);
                     }
