+++ src/main/java/com/amannmalik/mcp/util/ProgressCodec.java
@@ -35,7 +35,7 @@ public final class ProgressCodec {
     private static ProgressToken toToken(JsonValue value) {
         return switch (value.getValueType()) {
             case STRING -> new ProgressToken.StringToken(InputSanitizer.requireClean(((JsonString) value).getString()));
-            case NUMBER -> new ProgressToken.NumericToken(((JsonNumber) value).doubleValue());
+            case NUMBER -> new ProgressToken.NumericToken(((JsonNumber) value).longValue());
             default -> throw new IllegalArgumentException("Invalid token type");
         };
     }
